Refactor to decimal for currency, add config & UI improvements
- Switched all currency/Bitcoin calculations to decimal for accuracy
- Added StringHelper.Truncate and applied to BitcoinNews.razor
- Updated SatsConverter.razor for decimals, loading state, and null checks
- Refactored ContactController to use IHttpClientFactory and TelegramOptions
- Improved PaymentManagerController API binding and inheritance
- Introduced IBitcoinExchangeRateProvider; updated all providers
- Refactored ExchangeRateProvider for DI and decimal averaging
- Enhanced App.razor with SEO meta tags, lang attr, and moved JS to app.js
- Wrapped MainLayout.razor content in ErrorBoundary, improved accessibility
- Registered new options classes and providers in Program.cs
- Added AppOptions and TelegramOptions for configuration
- Moved all JS to new app.js file
- Various UI and accessibility enhancements throughout

+++ src/FreedomBlaze/Models/BitcoinExchangeRateModel.cs
@@ -3,6 +3,6 @@
 public class BitcoinExchangeRateModel
 {
     public string ExchangeName { get; set; }
-    public double BitcoinRateInUSD { get; set; }
+    public decimal BitcoinRateInUSD { get; set; }
     public CurrencyExchangeRateModel CurrencyExchangeRate { get; set; }
 }
